feat: `cl-log` -> `clog` 서비스명 변경 대응
