--- cl-log-api/src/main/kotlin/org/depromeet/cllog/server/api/configuration/ApiConstants.kt
@@ -1,8 +1,0 @@
-package org.depromeet.cllog.server.api.configuration
-
-object ApiConstants {
-
-    private const val API_BASE_PATH_PREFIX = "/api"
-
-    const val API_BASE_PATH_V1 = "$API_BASE_PATH_PREFIX/v1"
-}
