--- cl-log-domain/build.gradle.kts
@@ -1,13 +1,0 @@
-dependencies {
-    implementation(project(":cl-log-global-utils"))
-    implementation(project(":cl-log-infrastructure"))
-}
-
-tasks {
-    jar {
-        enabled = true
-    }
-    bootJar {
-        enabled = false
-    }
-}
