+++ clog-api/build.gradle.kts
@@ -1,0 +1,15 @@
+dependencies {
+    implementation(project(":clog-global-utils"))
+    implementation(project(":clog-domain"))
+
+    implementation("org.springframework.boot:spring-boot-starter-web")
+}
+
+tasks {
+    jar {
+        enabled = false
+    }
+    bootJar {
+        enabled = true
+    }
+}
