+++ clog-api/src/main/kotlin/org/depromeet/clog/server/api/configuration/ApiConstants.kt
@@ -1,0 +1,8 @@
+package org.depromeet.clog.server.api.configuration
+
+object ApiConstants {
+
+    private const val API_BASE_PATH_PREFIX = "/api"
+
+    const val API_BASE_PATH_V1 = "$API_BASE_PATH_PREFIX/v1"
+}
