+++ clog-api/src/main/resources/banner.txt
