+++ clog-domain/build.gradle.kts
@@ -1,0 +1,13 @@
+dependencies {
+    implementation(project(":clog-global-utils"))
+    implementation(project(":clog-infrastructure"))
+}
+
+tasks {
+    jar {
+        enabled = true
+    }
+    bootJar {
+        enabled = false
+    }
+}
